third_party/blink/web_tests/platform/generic/vibration/cancel-vibration-during-pattern-vibrating-expected.txt

Tests the cancelVibration during pattern vibration is working.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.isVibrating(navigator) is false
PASS internals.isVibrating(navigator) is true
PASS internals.isVibrating(navigator) is false
PASS internals.isVibrating(navigator) is true
PASS internals.isVibrating(navigator) is false
PASS internals.isVibrating(navigator) is true
PASS internals.isVibrating(navigator) is false
PASS internals.isVibrating(navigator) is true
PASS internals.isVibrating(navigator) is false
PASS successfullyParsed is true

TEST COMPLETE
